+++ tests/test_complete_files.py
@@ -169,6 +169,13 @@ class TestCompleteFiles(unittest.TestCase):
         # Exemple taken from svg of the flag of the state of California
         filename = "R_letter_with_arc"
         create_test_from_filename(filename, self)
+        c
+
+    def test_arc_direction(self):
+        """Test svg with S command inside it"""
+        # Exemple taken from svg of the flag of the state of California
+        filename = "s_command_letter"
+        create_test_from_filename(filename, self)
 
 
 if __name__ == "__main__":